Now the plants are stored localy in the user phone local storage
I  use now the fetch and save data method in the correct class
so now when the user create a plant it is storage in the local storage
and now the main class verify if the user have any plants so he can be
send to the add plant activity automatically. I add a method to the
UserPlant class so when the plant its created the UserPlant class store
the imageId of the plant in its own variable.

The button of info of the top corner stop working, to fixed i new to
modify the class of UserPlant